[system] Remove files that should be local-only
This is a result of a change in the way I use this repo:
Before, the $XDG_CONFIG_HOME directory was just simlinked to this.
Now, I have a dedicated config folder and when I feel like there is
a file/folder that I should backup/share, I move it to the repo
and create a new symlink.